[resolves #93] Use a unique Java package in example-camel-itests
