+++ itests/src/main/java/org/wildfly/camel/examples/test/common/FileConsumerTestSupport.java
@@ -1,0 +1,35 @@
+/*
+ * #%L
+ * Wildfly Camel :: Testsuite
+ * %%
+ * Copyright (C) 2013 - 2016 RedHat
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package org.wildfly.camel.examples.test.common;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public abstract class FileConsumerTestSupport {
+
+    @Test
+    public void testFileProcessed() throws Exception {
+        boolean logMessagePresent = ServerLogReader.awaitLogMessage(getExpectedLogMessage(), 10000);
+        Assert.assertTrue(logMessagePresent);
+    }
+
+    protected abstract String getExpectedLogMessage();
+    protected abstract String getContextName();
+}
